+++ CHANGELOG.md
@@ -1,13 +1,34 @@
-![Changelog v1.0.0](https://img.shields.io/badge/CHANGELOG-v1.0.0-orange) 
+![Changelog v1.1.0](https://img.shields.io/badge/CHANGELOG-v1.1.0-orange)
+
 # CHANGELOG
+
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- work on new features
+
+### Added
+
+- Action `apt_packages` updates package information and
+  installs any missing, required packages with `apt`.
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
 
 ## [0.0.1] - 2025-05-23
+
 ### Added
+
 - create install_dotfiles with [pforret/bashew](https://github.com/pforret/bashew)
+
+[unreleased]: https://github.com/jfishe/install_dotfiles/compare/v0.0.1...HEAD
+[0.0.1]: https://github.com/jfishe/install_dotfiles/releases/tag/v0.0.1